Change object positions to double (not float), and a plan for level format

+++ github.Game/src/net/earthcomputer/githubgame/geom/Pos.java
@@ -7,13 +7,13 @@ package net.earthcomputer.githubgame.geom;
  *
  */
 public class Pos {
-	private float x;
-	private float y;
+	private double x;
+	private double y;
 
 	/**
 	 * Constructs a position from x- and y-co-ordinates
 	 */
-	public Pos(float x, float y) {
+	public Pos(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -21,28 +21,28 @@ public class Pos {
 	/**
 	 * Gets the x-position of this position
 	 */
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
 	/**
 	 * Gets the y-position of this position
 	 */
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
 	/**
 	 * Sets the x-position of this position
 	 */
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
 	/**
 	 * Sets the y-position of this position
 	 */
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
@@ -57,7 +57,7 @@ public class Pos {
 	/**
 	 * Adds the given co-ordinates to the co-ordinates of this position
 	 */
-	public Pos add(float x, float y) {
+	public Pos add(double x, double y) {
 		this.x += x;
 		this.y += y;
 		return this;
@@ -84,17 +84,17 @@ public class Pos {
 		}
 
 		@Override
-		public void setX(float x) {
+		public void setX(double x) {
 			throw new UnsupportedOperationException("Cannot set the x-position of an unmodifiable position");
 		}
 
 		@Override
-		public void setY(float y) {
+		public void setY(double y) {
 			throw new UnsupportedOperationException("Cannot set the y-position of an unmodifiable position");
 		}
 
 		@Override
-		public Pos add(float x, float y) {
+		public Pos add(double x, double y) {
 			throw new UnsupportedOperationException("Cannot change the position of an unmodifiable position");
 		}
